refactor:  separate init logic
